correction 2 of Q1

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,5 @@
 #include "function_pointers.h"
+#include <stdlib.h>
 
 /**
  * array_iterator - iterates through an array
@@ -11,7 +12,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	size_t i;
 
 	for (i = 0; i < size; i++)
 		(*action)(array[i]);
